+++ pkg/netcalc/network.go
@@ -1,11 +1,9 @@
-// Package network provides types and methods for working with
-// TCP/IP networks. Each network has an Address (uint32) and Mask (uint32)
-// that represent a give IPv4 network.
 package netcalc
 
 import (
 	"encoding/json"
 	"fmt"
+	"sort"
 )
 
 type Network struct {
@@ -77,4 +75,32 @@ func (n *Network) BroadcastAddress() uint32 {
 
 func (n *Network) HostCount() int {
 	return int(n.BroadcastAddress() - n.Address - 1)
+}
+
+// SummarizeNetworks takes an array of Network objects and returns a summarized Network.
+// Summarization occurs by sorting the array and then finding a common bit boundary that
+// a common mask and common network bits can be calculated.
+func SummarizeNetworks(networks []*Network) (*Network, error) {
+
+	if len(networks) == 0 {
+		return nil, fmt.Errorf("SummarizeNetwork: no networks to summarize")
+	} else if len(networks) == 1 {
+		return networks[0], nil
+	}
+
+	sort.Slice(networks, func(i, j int) bool {
+		return networks[i].Address < networks[j].Address
+	})
+
+	commonBits := networks[0].Address
+	var commonMask uint32
+
+	for idx := 1; idx < len(networks); idx++ {
+		commonBits = commonBits & networks[idx].Address
+		commonMask = GetCommonBitMask(commonBits, networks[idx].Address)
+		commonBits = GetNetworkAddress(commonBits, commonMask)
+	}
+
+	return &Network{Address: commonBits, Mask: commonMask}, nil
+
 }
